requirements.txt
opencv-python
tqdm
tensorflow-gpu==1.14
keras==2.0.8
scikit-image
numpy